requirements.txt
requests>=2.25.1
python-dotenv>=0.19.0
matplotlib>=3.5.0
numpy>=1.21.0
tqdm>=4.64.0seaborn
